+++ lib/davinci_pas_test_kit/custom_groups/v2.0.1/client_tests/pas_client_pas_inquire_request_bundle_validation_test.rb
